Some more fixes
